requirements.txt
django~=3.2.5
django-filter~=2.4.0
djangorestframework~=3.12.0
djangorestframework-gis~=0.17
psycopg2~=2.9.0
factory_boy==3.2.0
django-cors-headers==3.7.0
djangorestframework==3.12.4
django_celery_beat==2.2.0